+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Do not allow negative numbers for most number input fields
 * Improve clarity for Lacks Bridge and Overcrowded details in settlement dialog 
+* Do not apply subsist after exploration penalty in macro if unchecked
 
 ### Changed